+++ tests/stores/data/__init__.py
@@ -1,1 +1,2 @@
+__test__ = False
 from .basicwords import *